\name{rcpp11test-package}
\alias{rcpp11test-package}
\alias{rcpp11test}
\docType{package}
\title{
\packageTitle{rcpp11test}
}
\description{
\packageDescription{rcpp11test}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rcpp11test}
\packageIndices{rcpp11test}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rcpp11test}

Maintainer: \packageMaintainer{rcpp11test}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
